Refactor InteractiveTutorial to use useSpotlight hook, add unit tests, and resolve layered architecture db import warning
